Fix bug

+++ src/applications/microbenchmark.h
@@ -65,7 +65,7 @@ class Microbenchmark : public Application {
   inline int RandomLocalKey(const int key_start, const int key_limit, const int part, Rand* rand) const {
 		return key_start + part + nparts * (rand->next() % ((key_limit - key_start)/nparts));
   }
-  inline int NotSoRandomLocalKey(const int rand_num, const int key_start, const int key_limit, const int part) const {
+  inline int NotSoRandomLocalKey(const int64 rand_num, const int key_start, const int key_limit, const int part) const {
 		return key_start + part + nparts * (abs(rand_num) % ((key_limit - key_start)/nparts));
   }
   void GetKeys(TxnProto* txn) const;
